refactor(graph): move the viewport translation members to PanningMixin
AbstractGraph still held members that belong to the panning concern and have no reason to sit in the class. They all
adjust the viewport without touching the scale, which is what PanningMixin already does with scrollCellToVisible and
scrollRectToVisible.

Move scrollPointToVisible, center, and the ignoreScrollbars and translateToScrollPosition flags with their getters.
scrollPointToVisible reads isTimerAutoScroll, isAllowAutoPanning, getPanDx, getPanDy and the PanningHandler plugin,
which PanningMixin owns entirely, and PanningMixin.type.ts already referenced the method as if it lived there.

center is included on purpose although FitPlugin looks like the obvious target. It never touches view.scale, it only
updates view.translate and the container scroll position, and it branches on hasScrollbars exactly like
scrollRectToVisible. Scale is the dividing line in this codebase, ZoomMixin and FitPlugin own it while PanningMixin owns
translate and scroll offset, so center belongs here. The name is what makes FitPlugin look right: fitCenter centers as a
consequence of fitting, deriving the translation from a newly computed scale, whereas center works at the current scale.

The public API is unchanged, mixin members are merged into the AbstractGraph interface by declaration merging.

One nuance for consumers: isIgnoreScrollbars and isTranslateToScrollPosition were arrow-function class properties, so
they were bound to their instance and could be detached. They are now prototype methods, like every other mixin member,
so `const f = graph.isIgnoreScrollbars; f()` no longer works and needs a call on the graph, or an explicit bind.
